Move example model migration to roadsign plugin
