+++ package.json
@@ -112,7 +112,7 @@
     "@glint/template": "^1.5.0",
     "@graphy/content.ttl.write": "^4.3.7",
     "@graphy/memory.dataset.fast": "4.3.3",
-    "@lblod/ember-rdfa-editor": "12.12.1-dev.cfccfaf88a64bd52aea209b09d95c4e705c0ca44",
+    "@lblod/ember-rdfa-editor": "12.12.1-dev.ecda548f934ee5eb8f0aeb3dd44290ace5c25621",
     "@rdfjs/types": "^1.1.0",
     "@release-it/keep-a-changelog": "^4.0.0",
     "@tsconfig/ember": "^3.0.8",
